+++ src/ethereum_test_specs/blockchain.py
@@ -364,7 +364,7 @@ class BlockchainTest(BaseTest):
             logs_bloom=0,
             difficulty=0x20000 if env.difficulty is None else env.difficulty,
             number=0,
-            gas_limit=env.gas_limit,
+            gas_limit=env.block_gas_limit,
             gas_used=0,
             timestamp=0,
             extra_data=b"\x00",